# static/specs/openapi/catalog/catalog_agent.yaml
openapi: 3.0.3
info:
  title: Rainbow Catalog Agent API - Entity Management
  description: OpenAPI specification for the Rainbow Catalog Agent Entity Management endpoints.
  version: 1.0.0
servers:
  - url: /v1/catalog-agent
    description: Base path for the catalog agent API
...
security:
  - BearerAuth: []
paths:
...
  /data-services:
    get:
      tags:
        - Data Services
      summary: Get all data services
      operationId: getAllDataServices
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: List of data services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataServiceDto"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Data Services
      summary: Create data service
      operationId: createDataService
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewDataServiceDto"
            example:
              dcatEndpointUrl: "http://example.com/api"
              catalogId: "urn:uuid:my-catalog-id"
              dctTitle: "My Data Service"
      responses:
        "200":
          description: Data service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataServiceDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /data-services/catalog/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Data Services
      summary: Get data services by catalog ID
      operationId: getDataServicesByCatalogId
      responses:
        "200":
          description: List of data services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataServiceDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /data-services/batch:
    post:
      tags:
        - Data Services
      summary: Get batch data services
      operationId: getBatchDataServices
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRequests"
      responses:
        "200":
          description: List of data services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataServiceDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /data-services/main:
    get:
      tags:
        - Data Services
      summary: Get main data service
      operationId: getMainDataService
      responses:
        "200":
          description: Main data service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataServiceDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Data Services
      summary: Create main data service
      operationId: createMainDataService
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewDataServiceDto"
      responses:
        "200":
          description: Main data service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataServiceDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /data-services/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Data Services
      summary: Get data service by ID
      operationId: getDataServiceById
      responses:
        "200":
          description: Data service found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataServiceDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Data Services
      summary: Update data service by ID
      operationId: updateDataServiceById
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditDataServiceDto"
      responses:
        "200":
          description: Data service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataServiceDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Data Services
      summary: Delete data service by ID
      operationId: deleteDataServiceById
      responses:
        "202":
          description: Data service deleted (Accepted)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --- Datasets ---
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        format: int64
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        format: int64
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
  schemas:
...
    ErrorInfo:
      type: object
      required:
        - message
        - error_code
        - cause
      properties:
        message:
          type: string
        error_code:
          type: integer
        details:
          type: string
        cause:
          type: string
      example:
        message: "Resource not found"
        error_code: 3200
        status_code: 404
        cause: "Catalog with ID urn:uuid:123 not found"

    # --- Common Wrappers ---
    BatchRequests:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
            format: urn
...
    DataServiceDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        dcatEndpointDescription:
          type: string
        dcatEndpointUrl:
          type: string
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctIdentifier:
          type: string
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dctDescription:
          type: string
        catalogId:
          type: string
          format: urn
        dspaceMainDataService:
          type: boolean
      example:
        {
          "catalogId": "urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635",
          "dcatEndpointDescription": null,
          "dcatEndpointUrl": "http://127.0.0.1:1200/dsp/current",
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctIdentifier": "urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f",
          "dctIssued": "2026-01-09T10:49:53.895893Z",
          "dctModified": null,
          "dctTitle": null,
          "dspaceMainDataService": true,
          "id": "urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f"
        }
    NewDataServiceDto:
      type: object
      required:
        - dcatEndpointUrl
        - catalogId
      properties:
        id:
          type: string
          format: urn
        dcatEndpointDescription:
          type: string
        dcatEndpointUrl:
          type: string
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctTitle:
          type: string
        dctDescription:
          type: string
        catalogId:
          type: string
          format: urn
      example:
        {
          "catalogId": "urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635",
          "dcatEndpointDescription": null,
          "dcatEndpointUrl": "http://127.0.0.1:1200/dsp/current",
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctTitle": null
        }
    EditDataServiceDto:
      type: object
      properties:
        dcatEndpointDescription:
          type: string
        dcatEndpointUrl:
          type: string
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctTitle:
          type: string
        dctDescription:
          type: string
      example:
        {
          "dcatEndpointDescription": null,
          "dcatEndpointUrl": "http://127.0.0.1:1200/dsp/current",
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctTitle": null
        }

    # --- Datasets ---
...